+++ src/Akeneo/Pim/Enrichment/Component/Product/Completeness/MissingRequiredAttributesCalculator.php
@@ -7,8 +7,6 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Completeness;
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodesCollection;
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Query\GetCompletenessProductMasks;
 use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithFamilyInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Structure\Component\Query\PublicApi\Family\GetRequiredAttributesMasks;
 
 /**
@@ -42,26 +40,15 @@ class MissingRequiredAttributesCalculator
     public function fromEntityWithFamily(
         EntityWithFamilyInterface $entityWithFamily
     ): ProductCompletenessWithMissingAttributeCodesCollection {
-        if (!$entityWithFamily instanceof ProductInterface && !$entityWithFamily instanceof ProductModelInterface) {
-            throw new \InvalidArgumentException(sprintf(
-                'Entity must be a product or a product model, instance of \'%s\' given',
-                get_class($entityWithFamily)
-            ));
-        }
-
         if (null === $entityWithFamily->getFamily()) {
             return new ProductCompletenessWithMissingAttributeCodesCollection($entityWithFamily->getId(), []);
         }
         $familyCode = $entityWithFamily->getFamily()->getCode();
         $requiredAttributesMasks = $this->getRequiredAttributesMasks->fromFamilyCodes([$familyCode]);
 
-        $identifier = $entityWithFamily instanceof ProductInterface ?
-            $entityWithFamily->getIdentifier() :
-            $entityWithFamily->getCode();
-
         $productMask = $this->getCompletenessProductMasks->fromValueCollection(
             $entityWithFamily->getId(),
-            $identifier,
+            $entityWithFamily->getIdentifier(),
             $familyCode,
             $entityWithFamily->getValues()
         );