+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/Normalization/ProductProcessor.php
@@ -4,7 +4,6 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Normalizat
 
 use Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\FilterValues;
 use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithFamilyInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\ValuesFiller\FillMissingValuesInterface;
 use Akeneo\Pim\Structure\Component\Repository\AttributeRepositoryInterface;
@@ -121,15 +120,7 @@ class ProductProcessor implements ItemProcessorInterface, StepExecutionAwareInte
      */
     protected function fetchMedia(EntityWithFamilyInterface $product, $directory)
     {
-        if (!$product instanceof ProductInterface && !$product instanceof ProductModelInterface) {
-            throw new \InvalidArgumentException(sprintf(
-                'Entity must be a product or a product model, instance of \'%s\' given',
-                get_class($product)
-            ));
-        }
-
-        $identifier = $product instanceof ProductModelInterface ? $product->getCode() : $product->getIdentifier();
-        $this->mediaFetcher->fetchAll($product->getValues(), $directory, $identifier);
+        $this->mediaFetcher->fetchAll($product->getValues(), $directory, $product->getIdentifier());
 
         foreach ($this->mediaFetcher->getErrors() as $error) {
             $this->stepExecution->addWarning($error['message'], [], new DataInvalidItem($error['media']));