+++ src/Akeneo/Pim/Enrichment/Component/Product/Validator/Constraints/ImmutableVariantAxesValuesValidator.php
@@ -6,8 +6,7 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Validator\Constraints;
 
 use Akeneo\Pim\Enrichment\Component\Product\EntityWithFamilyVariant\EntityWithFamilyVariantAttributesProvider;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\WriteValueCollectionFactory;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithFamilyVariantInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -45,11 +44,8 @@ class ImmutableVariantAxesValuesValidator extends ConstraintValidator
      */
     public function validate($entity, Constraint $constraint): void
     {
-        if (!$entity instanceof ProductInterface && !$entity instanceof ProductModelInterface) {
-            throw new \InvalidArgumentException(sprintf(
-                'Entity must be a product or a product model, instance of \'%s\' given',
-                get_class($entity)
-            ));
+        if (!$entity instanceof EntityWithFamilyVariantInterface) {
+            throw new UnexpectedTypeException($constraint, EntityWithFamilyVariantInterface::class);
         }
 
         if (!$constraint instanceof ImmutableVariantAxesValues) {