+++ src/Akeneo/Pim/Enrichment/Component/Product/Validator/Constraints/UniqueVariantAxisValidator.php
@@ -6,7 +6,6 @@ use Akeneo\Pim\Enrichment\Component\Product\EntityWithFamilyVariant\EntityWithFa
 use Akeneo\Pim\Enrichment\Component\Product\Exception\AlreadyExistingAxisValueCombinationException;
 use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithFamilyVariantInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\WriteValueCollection;
 use Akeneo\Pim\Enrichment\Component\Product\ProductModel\Query\GetValuesOfSiblings;
 use Akeneo\Pim\Enrichment\Component\Product\Validator\UniqueAxesCombinationSet;
@@ -209,20 +208,8 @@ class UniqueVariantAxisValidator extends ConstraintValidator
         $this->context->buildViolation($message, [
             '%values%' => $combination,
             '%attributes%' => $axesCodes,
-            '%validated_entity%' => $this->getEntityIdentifier($entityWithFamilyVariant),
+            '%validated_entity%' => $entityWithFamilyVariant->getIdentifier(),
             '%sibling_with_same_value%' => $siblingIdentifier,
         ])->atPath('attribute')->addViolation();
-    }
-
-    private function getEntityIdentifier(EntityWithFamilyVariantInterface $entity): string
-    {
-        if (!$entity instanceof ProductInterface && !$entity instanceof ProductModelInterface) {
-            throw new \InvalidArgumentException(sprintf(
-                'Entity must be a product or a product model, instance of \'%s\' given',
-                get_class($entity)
-            ));
-        }
-
-        return $entity instanceof ProductInterface ? $entity->getIdentifier() : $entity->getCode();
     }
 }